config/requirements.txt
agno
pypdf
openai
anthropic
google-generativeai
reportlab
lancedb

# Email processing dependencies
python-dotenv

# Additional utilities for email agent
schedule
psutil
